requirement_short.txt
torch==1.8.1
torchvision==0.9.1
pillow==10.0.1
numpy==1.26.0
matplotlib==3.8.0
e2cnn==0.2.2
einops==0.6.0